LMSA-9275 - Move the rest controller into the correct package so that all the swagger stuff works
